refactor(app): extract useQuickBattleFlow (#316 phase 9b)
Every non-campaign route into a battle: quick battle, card draft, endless,
the daily and weekly challenges, training, and the post-battle play-again
branch (world battle / training / draft / endless / quick).

Takes its 13 battle-mode refs as explicit arguments rather than bundling them.
Grouping them behind one object would have meant rewriting ~190 call sites
still in App for no behavioural gain, and the reset sets are deliberately not
uniform — every launch path sets battleFlawlessRef true except training, which
sets it false — so a shared "reset for new battle" helper would quietly change
behaviour. Documented on the args interface so the next reader doesn't
"simplify" it back.

App.tsx: 2,543 -> 2,321 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
